+++ cf/data/data.py
@@ -414,6 +414,8 @@ place.
             *Parameter example:*
                 ``dtype=numpy.dtype('i2')``
 
+            .. versionaddedd:: 3.0.4
+
         mask: optional
             Apply this mask to the data given by the *array*
             parameter. By default, or if *mask* is `None`, no mask is
@@ -424,6 +426,8 @@ place.
 
             This mask will applied in addition to any mask already
             defined by the *array* parameter.
+
+            .. versionaddedd:: 3.0.5
 
         source: optional
             Initialize the array, units, calendar and fill value from